requirements.txt
openai>=1.6.0
pyyaml>=6.0
python-dotenv>=1.0.0
streamlit>=1.22.0
regex>=2023.0.0
typing-extensions>=4.5.0 